fix(schemas/properties): change the way to call properties of schemas
