+++ packages/api-gateway/src/drivers/http/routes/usuarios/schema/createUsuarioSchema.js
@@ -1,6 +1,6 @@
-const { usuario, persona, general } = require('./generalSchema');
-
-const personaResponseSchema = require('./personaResponseSchema');
+const { persona } = require('./properties/persona');
+const { responseProperties } = require('./properties/responseProperties');
+const { usuario } = require('./properties/usuario');
 
 const createUsuarioSchema = {
   tags: ['Usuario'],
@@ -8,32 +8,17 @@ const createUsuarioSchema = {
   body: {
     type: 'object',
     properties: {
-      rolId: usuario.rolId,
-      usuario: usuario.usuario,
-      correo: usuario.correo,
-      contrasena: usuario.contrasena,
-      estatus: usuario.estatus,
-      actualizado: usuario.actualizado,
+      ...usuario,
+      contrasena: { type: 'string', minLength: 3, maxLength: 25 },
       persona: {
         type: 'object',
         properties: {
-          nombre: persona.nombre,
-          apellidoPaterno: persona.apellidoPaterno,
-          apellidoMaterno: persona.apellidoMaterno,
-          fechaNacimiento: persona.fechaNacimiento,
-          sexo: persona.sexo,
-          nacionalidad: persona.nacionalidad,
-          telefono: persona.telefono,
-          celular: persona.celular,
-          curp: persona.curp,
-          rfc: persona.rfc,
-          ine: persona.ine,
-          fotografia: persona.fotografia,
+          ...persona,
         },
         required: ['nombre', 'apellidoPaterno'],
       },
     },
-    required: ['roldId', 'usuario', 'correo', 'contrasena', 'actualizado'],
+    required: ['rolId', 'usuario', 'correo', 'contrasena', 'actualizado'],
   },
   response: {
     201: {
@@ -42,14 +27,17 @@ const createUsuarioSchema = {
         data: {
           type: 'object',
           properties: {
-            id: usuario.id,
-            rolId: usuario.rolId,
-            usuario: usuario.usuario,
-            correo: usuario.correo,
-            estatus: usuario.estatus,
-            createdAt: general.createdAt,
-            updatedAt: general.updatedAt,
-            persona: personaResponseSchema,
+            id: { type: 'integer' },
+            ...usuario,
+            ...responseProperties,
+            persona: {
+              type: 'object',
+              properties: {
+                id: { type: 'integer' },
+                ...persona,
+                ...responseProperties,
+              },
+            },
           },
         },
       },
