+++ packages/api-gateway/src/drivers/http/routes/usuarios/schema/deleteUsuarioSchema.js
@@ -1,8 +1,6 @@
-const {
-  usuario,
-  general,
-} = require('./generalSchema');
-const personaResponseSchema = require('./personaResponseSchema');
+const { persona } = require('./properties/persona');
+const { responseProperties } = require('./properties/responseProperties');
+const { usuario } = require('./properties/usuario');
 
 const deleteUsuarioSchema = {
   tags: ['Usuario'],
@@ -10,7 +8,7 @@ const deleteUsuarioSchema = {
   params: {
     type: 'object',
     properties: {
-      usuarioId: usuario.usuarioId,
+      usuarioId: { type: 'integer' },
     },
     required: ['usuarioId'],
   },
@@ -21,15 +19,17 @@ const deleteUsuarioSchema = {
         data: {
           type: 'object',
           properties: {
-            id: usuario.id,
-            rolId: usuario.rolId,
-            usuario: usuario.usuario,
-            correo: usuario.correo,
-            estatus: usuario.estatus,
-            createdAt: general.createdAt,
-            updatedAt: general.updatedAt,
-            deletedAt: general.deletedAt,
-            persona: personaResponseSchema,
+            id: { type: 'integer' },
+            ...usuario,
+            ...responseProperties,
+            persona: {
+              type: 'object',
+              properties: {
+                id: { type: 'integer' },
+                ...persona,
+                ...responseProperties,
+              },
+            },
           },
         },
       },
