+++ packages/api-gateway/src/drivers/http/routes/usuarios/schema/properties/domicilio.js
@@ -1,0 +1,14 @@
+// Domicilio
+const domicilio = {
+  municipioId: { type: 'integer' },
+  estadoId: { type: 'integer' },
+  calle: { type: 'string' },
+  numeroExterior: { type: 'string' },
+  numeroInterior: { type: 'string' },
+  colonia: { type: 'string' },
+  codigoPostal: { type: 'integer' },
+  latitud: { type: 'string' },
+  longitud: { type: 'string' },
+};
+
+module.exports = { domicilio };
