+++ packages/api-gateway/src/drivers/http/routes/usuarios/schema/properties/estado.js
@@ -1,0 +1,7 @@
+// Estado
+const estado = {
+  paisId: { type: 'integer' },
+  estado: { type: 'string' },
+};
+
+module.exports = { estado };
