+++ packages/api-gateway/src/drivers/http/routes/usuarios/schema/properties/municipio.js
@@ -1,0 +1,7 @@
+// Municipio
+const municipio = {
+  estadoId: { type: 'integer' },
+  municipio: { type: 'string' },
+};
+
+module.exports = { municipio };
