+++ packages/api-gateway/src/drivers/http/routes/usuarios/schema/properties/responseProperties.js
@@ -1,0 +1,8 @@
+// General
+const responseProperties = {
+  createdAt: { type: 'string', format: 'date-time' },
+  updatedAt: { type: 'string', format: 'date-time' },
+  deletedAt: { type: 'string', format: 'date-time' },
+};
+
+module.exports = { responseProperties };
